requirements.txt
telethon
python-dotenv
asyncpg
httpx
colorama